sample(app): move auth demos into auth subpackage and add firestore emulator toggle
