+++ app/src/main/java/com/firebaseui/android/demo/MainActivity.kt
@@ -31,12 +31,18 @@ import com.firebaseui.android.demo.database.DatabaseDemoActivity
 import com.firebaseui.android.demo.firestore.FirestoreDemoActivity
 import com.firebaseui.android.demo.storage.StorageDemoActivity
 import com.google.firebase.FirebaseApp
+import com.google.firebase.firestore.FirebaseFirestore
 
 class MainActivity : ComponentActivity() {
     companion object {
-        private const val USE_AUTH_EMULATOR = false
+        private const val USE_AUTH_EMULATOR = true
         private const val AUTH_EMULATOR_HOST = "10.0.2.2"
         private const val AUTH_EMULATOR_PORT = 9099
+
+        // 10.0.2.2 is the Android emulator's alias for the host machine's localhost.
+        private const val USE_FIRESTORE_EMULATOR = true
+        private const val FIRESTORE_EMULATOR_HOST = "10.0.2.2"
+        private const val FIRESTORE_EMULATOR_PORT = 8080
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -48,6 +54,11 @@ class MainActivity : ComponentActivity() {
 
         if (USE_AUTH_EMULATOR) {
             authUI.auth.useEmulator(AUTH_EMULATOR_HOST, AUTH_EMULATOR_PORT)
+        }
+
+        if (USE_FIRESTORE_EMULATOR) {
+            FirebaseFirestore.getInstance()
+                .useEmulator(FIRESTORE_EMULATOR_HOST, FIRESTORE_EMULATOR_PORT)
         }
 
         var pendingEmailLink = intent.getStringExtra(EmailLinkConstants.EXTRA_EMAIL_LINK)
