+++ e2eTest/firebase.json
@@ -3,6 +3,12 @@
     "auth": {
       "port": 9099
     },
+    "firestore": {
+      "port": 8080
+    },
+    "database": {
+      "port": 8199
+    },
     "ui": {
       "enabled": true
     },
